pyproject.toml
[project]
name = "miraita"
version = "0.1.0"
description = "基于 Entari 开发，超可爱的林汐酱"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arclet-entari[full,pydantic]>=0.17.0rc3",
    "entari-plugin-database>=0.2.1",
    "entari-plugin-server>=0.2.3",
    "fastapi>=0.116.1",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "py-cpuinfo>=9.0.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.7",
    "email-validator>=2.2.0",
    "satori-python-adapter-onebot11>=0.1.0",
    "satori-python-adapter-satori>=0.1.0",
    "rich>=14.1.0",
]
llm = [
    "docstring-parser>=0.17.0",
    "litellm>=1.81.14",
]
prometheus = [
    "prometheus-client>=0.22.1",
]

[tool.uv.build-backend]
module-root = ""

[tool.ruff]
line-length = 88

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.ruff.lint.isort]
length-sort = true
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"

typeCheckingMode = "standard"
reportPrivateImportUsage = false
reportShadowedImports = false
disableBytesTypePromotions = true

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"
